--- a/packages/falcon-client/src/i18n/i18nServerFactory.js
+++ b/packages/falcon-client/src/i18n/i18nServerFactory.js
@@ -47,10 +47,10 @@ export default ({ lng = 'en', fallbackLng = 'en', whitelist = ['en'], debug = fa
     );
   });
 
-export function extractI18nextState(ctx) {
-  if (ctx.i18next) {
+export function extractI18nextState(i18nextInstance) {
+  if (i18nextInstance) {
     return {
-      language: ctx.i18next.language
+      language: i18nextInstance.language
     };
   }