--- a/packages/falcon-client/src/i18n/i18nServerFactory.js
+++ b/packages/falcon-client/src/i18n/i18nServerFactory.js
@@ -1,5 +1,5 @@
 import i18next from 'i18next';
-import Backend from 'i18next-sync-fs-backend';
+import Backend from 'i18next-node-fs-backend';
 
 /**
  * @typedef {object} Options