--- a/packages/falcon-client/src/middlewares/routes/appHtmlMiddleware.js
+++ b/packages/falcon-client/src/middlewares/routes/appHtmlMiddleware.js
@@ -7,7 +7,7 @@ import Html from '../../components/Html';
  * Application html renderer middleware.
  * @param {Object} params params
  * @param {Object} params.config configuration
- * @returns {function(ctx: object, next: function): Promise<void>} Koa middleware
+ * @returns {import('koa').Middleware} Koa middleware
  */
 export default ({ config }) => async ctx => {
   const { AppMarkup, client, i18next, assets, chunkExtractor, helmetContext, serverTiming } = ctx.state;