--- a/packages/falcon-client/src/middlewares/routes/appShellMiddleware.js
+++ b/packages/falcon-client/src/middlewares/routes/appShellMiddleware.js
@@ -7,7 +7,7 @@ import HtmlHead from '../../components/HtmlHead';
 
 /**
  * App shell rendering middleware.
- * @returns {function(ctx: object, next: function): Promise<void>} Koa middleware
+ * @returns {import('koa').Middleware} Koa middleware
  */
 export default () => async (ctx, next) => {
   const { client, chunkExtractor } = ctx.state;