--- a/packages/falcon-client/src/middlewares/routes/index.js
+++ b/packages/falcon-client/src/middlewares/routes/index.js
@@ -48,7 +48,7 @@ export function renderAppShell({ config, webpackAssets }) {
  * @param {RenderApp} params params
  * @returns {function(ctx: object, next: function)[]} Koa middlewares
  */
-export function renderApp({ config, clientApolloSchema, App, webpackAssets }) {
+export async function renderApp({ config, clientApolloSchema, App, webpackAssets }) {
   const { i18n, serverSideRendering, apolloClient } = config;
   const configSchema = { data: { config } };
 
@@ -61,7 +61,7 @@ export function renderApp({ config, clientApolloSchema, App, webpackAssets }) {
         clientApolloSchema
       }
     }),
-    i18next({ ...i18n }),
+    await i18next({ ...i18n }),
     serverSideRendering ? ssr({ App }) : appShell(),
     appHtml({ config })
   ].filter(x => x);