--- a/packages/falcon-client/src/middlewares/routes/index.js
+++ b/packages/falcon-client/src/middlewares/routes/index.js
@@ -5,13 +5,6 @@ import appHtml from './appHtmlMiddleware';
 import i18next from './i18nextMiddleware';
 import assets from './assetsMiddleware';
 
-/** Koa middleware
- * @callback KoaMiddleware
- * @param {Object} ctx context
- * @param {function} next next middleware
- * @returns {Promise<void>}
- */
-
 /**
  * @typedef {object} RenderAppShell
  * @property {object} config App configuration