--- a/packages/falcon-ecommerce-uikit/src/Checkout/TotalRow.tsx
+++ b/packages/falcon-ecommerce-uikit/src/Checkout/TotalRow.tsx
@@ -1,15 +1,14 @@
 import React from 'react';
 import { Box, Text } from '@deity/falcon-ui';
-import { Price } from './../Locale';
+import { Price } from '../Locale';
 
 export type TotalRowProps = {
   title: string;
   value: number;
-  currency?: string;
 };
-export const TotalRow: React.SFC<TotalRowProps> = ({ title, value, currency, ...props }) => (
+export const TotalRow: React.SFC<TotalRowProps> = ({ title, value, ...props }) => (
   <Box display="flex" {...props}>
     <Text flex="1">{title}</Text>
-    <Price value={value} currency={currency} />
+    <Price value={value} />
   </Box>
 );