--- a/packages/falcon-ecommerce-uikit/src/Locale/DateFormat.tsx
+++ b/packages/falcon-ecommerce-uikit/src/Locale/DateFormat.tsx
@@ -1,20 +1,26 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import { themed, Text } from '@deity/falcon-ui';
 import { Locale, DateTimeFormatOptions } from './LocaleContext';
 
 type DateFormatProps = {
-  value: string;
-  locale?: string;
-  options?: DateTimeFormatOptions;
+  value: number | string | Date;
+  /** passing formatting options may not use memoized formatter, so the performance penalty could be paid */
+  formatOptions?: DateTimeFormatOptions;
 };
 
-const DateFormatInnerDOM: React.SFC<DateFormatProps> = ({ value, locale, options, ...rest }) => (
-  <Locale>{({ dateTimeFormat }) => <Text {...rest}>{dateTimeFormat(value, { ...options, locale })}</Text>}</Locale>
+const DateFormatInnerDOM: React.SFC<DateFormatProps> = ({ value, formatOptions, ...rest }) => (
+  <Locale>{({ dateTimeFormat }) => <Text {...rest}>{dateTimeFormat(value, formatOptions)}</Text>}</Locale>
 );
+DateFormatInnerDOM.propTypes = {
+  value: PropTypes.oneOfType([PropTypes.number, PropTypes.string, PropTypes.instanceOf(Date)]).isRequired
+};
 
 export const DateFormat = themed({
   tag: DateFormatInnerDOM,
-
+  defaultProps: {
+    formatOptions: undefined
+  },
   defaultTheme: {
     price: {
       display: 'block',