### Please check if the PR fulfills these requirements ###

- [x] The commit message follows our guidelines

**What kind of change does this PR introduce? (Bug fix, feature, docs update, ...)**
*about 13% better results in average*
introduced performance optimizations in:
* internationalization setup
* values formatting

### What is the current behavior? 1️⃣ 
```
 $ hey -n 2000 -c 50 http://localhost:3000

Summary:
  Total:        53.5346 secs
  Slowest:      1.8916 secs
  Fastest:      0.6407 secs
  Average:      1.3279 secs
  Requests/sec: 37.3590


Response time histogram:
  0.641 [1]     |
  0.766 [8]     |
  0.891 [3]     |
  1.016 [22]    |■
  1.141 [33]    |■
  1.266 [107]   |■■■
  1.391 [1605]  |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  1.516 [202]   |■■■■■
  1.641 [17]    |
  1.767 [1]     |
  1.892 [1]     |


Latency distribution:
  10% in 1.2766 secs
  25% in 1.3068 secs
  50% in 1.3399 secs
  75% in 1.3671 secs
  90% in 1.3922 secs
  95% in 1.4235 secs
  99% in 1.5042 secs

Details (average, fastest, slowest):
  DNS+dialup:   0.0001 secs, 0.6407 secs, 1.8916 secs
  DNS-lookup:   0.0001 secs, 0.0000 secs, 0.0057 secs
  req write:    0.0000 secs, 0.0000 secs, 0.0014 secs
  resp wait:    1.2208 secs, 0.5169 secs, 1.4819 secs
  resp read:    0.1068 secs, 0.0005 secs, 0.6519 secs

Status code distribution:
  [200] 2000 responses
```

### What is the new behavior? 1️⃣ 
```
$ hey -n 2000 -c 50 http://localhost:3000

Summary:
  Total:        46.6591 secs
  Slowest:      1.5916 secs
  Fastest:      0.5469 secs
  Average:      1.1620 secs
  Requests/sec: 42.8641


Response time histogram:
  0.547 [1]     |
  0.651 [0]     |
  0.756 [6]     |
  0.860 [58]    |■■
  0.965 [55]    |■■
  1.069 [118]   |■■■■■
  1.174 [946]   |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  1.278 [594]   |■■■■■■■■■■■■■■■■■■■■■■■■■
  1.383 [145]   |■■■■■■
  1.487 [61]    |■■■
  1.592 [16]    |■


Latency distribution:
  10% in 1.0383 secs
  25% in 1.1338 secs
  50% in 1.1606 secs
  75% in 1.1968 secs
  90% in 1.2851 secs
  95% in 1.3654 secs
  99% in 1.4832 secs

Details (average, fastest, slowest):
  DNS+dialup:   0.0001 secs, 0.5469 secs, 1.5916 secs
  DNS-lookup:   0.0001 secs, 0.0000 secs, 0.0053 secs
  req write:    0.0000 secs, 0.0000 secs, 0.0027 secs
  resp wait:    1.0556 secs, 0.4669 secs, 1.4831 secs
  resp read:    0.1062 secs, 0.0023 secs, 0.4681 secs

Status code distribution:
  [200] 2000 responses

```

1️⃣  tested locally